+++ wp-content/themes/accelerate/podcasting-platform/components/parser.js
@@ -1,0 +1,3 @@
+export const parser = new DOMParser();
+export const xmlParse = (text) => parser.parseFromString(text, 'text/xml');
+export const htmlParse = (text) => parser.parseFromString(text, 'text/html');
