+++ wp-content/themes/accelerate/podcasting-platform/components/player.js
@@ -1,0 +1,7 @@
+import {xmlParse, htmlParse} from './parser.js'
+import Podcast, { PODCAST_STATE } from './podcast.js'
+
+export default class Player {
+    constructor() {
+    }
+  }
