+++ wp-content/themes/accelerate/podcasting-platform/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "podcasting-platform",
+  "version": "1.0.0",
+  "description": "npm for webpack and babel",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "Federico Santana",
+  "license": "ISC",
+  "devDependencies": {
+    "webpack": "^5.28.0",
+    "webpack-cli": "^4.5.0"
+  }
+}
