+++ wp-content/themes/accelerate/podcasting-platform/webpack.config.js
@@ -1,0 +1,11 @@
+const path = require('path');
+
+module.exports = {
+  devtool: false,
+  mode: 'development',
+  entry: './index.js',
+  output: {
+    filename: 'bundle.js',
+    path: path.resolve(__dirname, 'assets'),
+  }
+};
